requirements.txt
typer
pdfplumber
rich
python-docx